requirements.txt
circuitsvis>=1.43.2
datasets>=2.18.0
einops>=0.7.0
matplotlib>=3.8.3
nnsight>=0.2.11
pandas>=2.2.1
plotly>=5.18.0
torch>=2.1.2
tqdm>=4.66.1
umap-learn>=0.5.6
zstandard>=0.22.0
pytest>=8.1.1
transformers>=4.39.3
wandb>=0.16.0
# Chess specific requirements below
python-chess==1.999
pydantic==2.6.4
jaxtyping==0.2.28
beartype == 0.14.1

# For OthelloGPT
transformer_lens==1.15.0